cloud_function/requirements.txt
gcloud 
google-cloud-storage
google.cloud